+++ app/urls.py
@@ -16,7 +16,6 @@ urlpatterns = [
     path('blog/<slug:slug>/', views.blog_entry, name='blog_entry'),
     path('archive/', views.archive, name='archive'),
     path('register_subscriber/', views.register_subscriber, name='register_subscriber'),
-
     path('sitemap.xml/', sitemap, {'sitemaps': sitemaps}, name='django.contrib.sitemaps.views.sitemap')
 
     # TODO path('robots.txt/', views.robots, name='robots')