+++ app/views.py
@@ -1,15 +1,17 @@
 from django.shortcuts import render, get_object_or_404
 from django.template import Context
-from django.http import HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseBadRequest
 from django.core.exceptions import ValidationError
 
 from .models import Blog, Subscriber
+from .forms import SignupForm
 
 """
 """
 def about(request):
     context = {}
     return render(request, 'app/about.html', context)
+
 
 """
 """
@@ -21,17 +23,22 @@ def blog(request):
 
     return render(request, 'app/blog.html', context)
 
+
 """
 """
 def blog_entry(request, slug):
     entry = get_object_or_404(Blog, slug=slug)
+    context = {
+        "entry": get_object_or_404(Blog, slug=slug)
+    }
 
     #   Increment rating on page view
     entry.rating +=1
     #   Write updated rating to DB
     entry.save()
 
-    return render(request, 'app/blog_entry.html', { 'entry':entry })
+    return render(request, 'app/blog_entry.html', context)
+
 
 """
 """
@@ -42,22 +49,16 @@ def archive(request):
     }
     return render(request, 'app/archive.html', context)
 
+
 """
 """
 def register_subscriber(request):
     if request.method == 'POST':
-        form = SignupForm(request.POST or None)
-
+        form = SignupForm(request.POST)
         if form.is_valid():
-            username = form.clean_username()
-            if Subscriber.objects.filter(username=username).exists():
-                raise ValidationError(('Username: ' + username + ' already exists!'), code='invalid')
-                #TODO Add visual feedback on INVALID input
-                return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
-
-            print("New user subscribed: %s!", username)
-            Subscriber.objects.create(username=username)
-            #TODO Add visual feedback on VALID input
-            return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
-    else:
-        print("GET request")
+            print('[SUCCESS] : New sub: %s' % form.data['username'])
+            return HttpResponse('')
+        else:
+            print('[FAIL] : Invalid sub: %s' % form.data['username'])
+            return HttpResponseBadRequest('')
+    return render(request, 'app/blog.html', {})